requirements.txt
pytorch-lightning
torchmetrics
torchtext
transformers
icecream
datasets
matplotlib
tensorboard
tokenizations